apps/streamlit_dbsql/requirements.txt
streamlit==1.39.0
databricks-sdk==0.33.0
pandas==2.2.3
